DataViews: Share focus-on-mount mapping between modal layouts
Extract the `useMapFocusOnMount` helper out of `dataviews-item-actions`
into `hooks/use-map-focus-on-mount` and reuse it from the DataForm panel
modal.

`PanelModal` previously dropped to Base UI's smart default after the
Dialog migration. The smart default is fine in many cases, but for a
field-edit popup we want the first input focused (matching the legacy
`Modal.focusOnMount: 'firstInputElement'` behaviour). Reusing the same
helper keeps both layouts in sync and avoids duplicating the
input-selector heuristic.

Also wraps the `PanelModal` body in `Dialog.Content` so long forms scroll
while the title and footer stay sticky, matching the action-modal
treatment.
